requirements.txt
flask==3.0.3
#fastapi==0.112.0
asyncpg==0.29.0
sqlalchemy==2.0.31
uvicorn==0.30.5
pydantic==2.8.0
python_dotenv==1.0.1